requirements.txt
python-dotenv>=0.21.0
openai>=1.0.0
pydantic>=2.0.0